dialogflow_automation/requirements.txt
google-cloud-dialogflow>=2.23.0
python-dotenv==1.0.0
